[refactor] CreateRepository, UpdateRepository ->ChallengeRepository 및 ImageRepository 분리
