+++ Routinus/Routinus/Application/Coordinator/CreateCoordinator.swift
@@ -20,11 +20,13 @@ final class CreateCoordinator: RoutinusCoordinator {
         let challengeCreateUsecase = ChallengeCreateUsecase(repository: repository)
         let challengeUpdateUsecase = ChallengeUpdateUsecase(repository: repository)
         let challengeFetchUsecase = ChallengeFetchUsecase(repository: repository)
+        let imageSaveUsecase = ImageSaveUsecase(repository: repository)
         // TODO: - 챌린지 관리 화면으로부터 challengeID 주입받기
         let createViewModel = CreateViewModel(challengeID: "24F3374C-BFA7-49B2-AE26-F887009DC3DA",
                                               challengeCreateUsecase: challengeCreateUsecase,
                                               challengeUpdateUsecase: challengeUpdateUsecase,
-                                              challengeFetchUsecase: challengeFetchUsecase)
+                                              challengeFetchUsecase: challengeFetchUsecase,
+                                              imageSaveUsecase: imageSaveUsecase)
         let createViewController = CreateViewController(with: createViewModel)
         createViewController.hidesBottomBarWhenPushed = true
         self.navigationController.pushViewController(createViewController, animated: true)
