+++ Routinus/Routinus/Data/ChallengeRepository.swift
@@ -8,6 +8,7 @@
 import Foundation
 
 import RoutinusDatabase
+import RoutinusImageManager
 
 protocol ChallengeRepository {
     func fetchRecommendChallenges() async -> [Challenge]
@@ -17,6 +18,16 @@ protocol ChallengeRepository {
     func fetchChallenges(by userID: String) async -> [Challenge]
     func fetchChallenge(challengeID: String,
                         completion: @escaping (Challenge) -> Void)
+    func save(challenge: Challenge,
+              imageURL: String,
+              thumbnailImageURL: String,
+              authExampleImageURL: String,
+              authExampleThumbnailImageURL: String)
+    func update(challenge: Challenge,
+                imageURL: String,
+                thumbnailImageURL: String,
+                authExampleImageURL: String,
+                authExampleThumbnailImageURL: String) async
 }
 
 extension RoutinusRepository: ChallengeRepository {
@@ -52,6 +63,60 @@ extension RoutinusRepository: ChallengeRepository {
         RoutinusDatabase.challenge(ownerID: ownerID,
                                    challengeID: challengeID) { dto in
             completion(Challenge(challengeDTO: dto))
+        }
+    }
+
+    func save(challenge: Challenge,
+              imageURL: String,
+              thumbnailImageURL: String,
+              authExampleImageURL: String,
+              authExampleThumbnailImageURL: String) {
+        guard let startDate = challenge.startDate?.toString(),
+              let endDate = challenge.endDate?.toString() else { return }
+
+        let challengeDTO = ChallengeDTO(id: challenge.challengeID,
+                                        title: challenge.title,
+                                        authMethod: challenge.authMethod,
+                                        categoryID: challenge.category.id,
+                                        week: challenge.week,
+                                        desc: challenge.introduction,
+                                        startDate: startDate,
+                                        endDate: endDate,
+                                        participantCount: 1,
+                                        ownerID: challenge.ownerID)
+
+        RoutinusDatabase.createChallenge(challenge: challengeDTO,
+                                         imageURL: imageURL,
+                                         thumbnailImageURL: thumbnailImageURL,
+                                         authExampleImageURL: authExampleImageURL,
+                                         authExampleThumbnailImageURL: authExampleThumbnailImageURL) {
+            RoutinusImageManager.removeTempCachedImages()
+        }
+    }
+
+    func update(challenge: Challenge,
+                imageURL: String,
+                thumbnailImageURL: String,
+                authExampleImageURL: String,
+                authExampleThumbnailImageURL: String) async {
+        guard let startDate = challenge.startDate?.toString(),
+              let endDate = challenge.endDate?.toString() else { return }
+        let challengeDTO = ChallengeDTO(id: challenge.challengeID,
+                                        title: challenge.title,
+                                        authMethod: challenge.authMethod,
+                                        categoryID: challenge.category.id,
+                                        week: challenge.week,
+                                        desc: challenge.introduction,
+                                        startDate: startDate,
+                                        endDate: endDate,
+                                        participantCount: challenge.participantCount,
+                                        ownerID: challenge.ownerID)
+        RoutinusDatabase.patchChallenge(challengeDTO: challengeDTO,
+                                        imageURL: imageURL,
+                                        thumbnailImageURL: thumbnailImageURL,
+                                        authExampleImageURL: authExampleImageURL,
+                                        authExampleThumbnailImageURL: authExampleThumbnailImageURL) {
+            RoutinusImageManager.removeTempCachedImages()
         }
     }
 }
