--- Routinus/Routinus/Data/CreateRepository.swift
@@ -1,60 +1,0 @@
-//
-//  CreateRepository.swift
-//  Routinus
-//
-//  Created by 백지현 on 2021/11/09.
-//
-
-import Foundation
-
-import RoutinusDatabase
-import RoutinusImageManager
-
-protocol CreateRepository {
-    func save(challenge: Challenge,
-              imageURL: String,
-              thumbnailImageURL: String,
-              authExampleImageURL: String,
-              authExampleThumbnailImageURL: String)
-    func saveImage(to directory: String,
-                   filename: String,
-                   data: Data?) -> String?
-}
-
-extension RoutinusRepository: CreateRepository {
-    func save(challenge: Challenge,
-              imageURL: String,
-              thumbnailImageURL: String,
-              authExampleImageURL: String,
-              authExampleThumbnailImageURL: String) {
-        guard let startDate = challenge.startDate?.toString(),
-              let endDate = challenge.endDate?.toString() else { return }
-
-        let challengeDTO = ChallengeDTO(id: challenge.challengeID,
-                                        title: challenge.title,
-                                        authMethod: challenge.authMethod,
-                                        categoryID: challenge.category.id,
-                                        week: challenge.week,
-                                        desc: challenge.introduction,
-                                        startDate: startDate,
-                                        endDate: endDate,
-                                        participantCount: 1,
-                                        ownerID: challenge.ownerID)
-
-        RoutinusDatabase.createChallenge(challenge: challengeDTO,
-                                         imageURL: imageURL,
-                                         thumbnailImageURL: thumbnailImageURL,
-                                         authExampleImageURL: authExampleImageURL,
-                                         authExampleThumbnailImageURL: authExampleThumbnailImageURL) {
-            RoutinusImageManager.removeTempCachedImages()
-        }
-    }
-
-    func saveImage(to directory: String,
-                   filename: String,
-                   data: Data?) -> String? {
-        return RoutinusImageManager.saveImage(to: directory,
-                                              filename: filename,
-                                              imageData: data)
-    }
-}
